Stage 2 fix: cast bf16-autocast scores to fp32 for the code-score scatter
scatter_add_ requires matching dtypes; under autocast the score head
emits bf16 while the aggregation buffer is fp32 — crashed T-learned at
step one (CPU smoke ran without autocast so it passed).

+++ train_stage2.py
@@ -163,7 +163,7 @@ class VocabNeocore(nn.Module):
                         .scatter_add_(1, codes, one)
                     present = counts > 0
                     cs = torch.zeros(B, self.k_codes, device=dev) \
-                        .scatter_add_(1, codes, scores.detach())
+                        .scatter_add_(1, codes, scores.detach().float())
                     cs = cs.masked_fill(~present | admitted_codes,
                                         float("-inf"))
                     k_eff = min(k_r, int((~admitted_codes & present)
